refactor: image 수정에 관한 핸들러 함수를 PostUploadPage -> ImageUploader 컴포넌트로 이동
